server_grpc/requirements.txt
grpcio
grpcio-tools
asyncio
jwt
datetime
uuid
pydantic
fastapi
mariadb